Update firstVersion
Everything for detecting the edges is correct. The avgWidth, without the if statements within the for loops on lines 404 -429, will give a width for PPM of 1 more than it should be and PWM with width ~ 1 less than it should be. PPM should be 10 width with the conditions we set, but is at 9.99